backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
python-jose[cryptography]
passlib[bcrypt]
openai
python-dotenv
